refactor: consolidate code in eexecution*.go files
- Merge `engine_test.go` and `execution_test.go` into a single test file
- Remove redundant test code and consolidate test logic
- Update function and method names for consistency
- Simplify transaction and block processing in test cases
- Remove unnecessary comments and debug logging
