--- engine_test.go
@@ -1,191 +1,0 @@
-package execution
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	ethTypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/rollkit/go-execution/types"
-	"github.com/stretchr/testify/require"
-)
-
-// createEthClient creates an Ethereum client for checking block information
-func createEthClient(t *testing.T) *ethclient.Client {
-	t.Helper()
-
-	// Use the same ETH URL as in the tests
-	ethClient, err := ethclient.Dial(TEST_ETH_URL)
-	require.NoError(t, err, "Failed to create Ethereum client")
-
-	return ethClient
-}
-
-// checkLatestBlock retrieves and returns the latest block height, hash, and transaction count using Ethereum API
-func checkLatestBlock(t *testing.T, ctx context.Context) (uint64, common.Hash, int) {
-	t.Helper()
-
-	// Create an Ethereum client
-	ethClient := createEthClient(t)
-	defer ethClient.Close()
-
-	// Get the latest block header
-	header, err := ethClient.HeaderByNumber(ctx, nil) // nil means latest block
-	if err != nil {
-		t.Logf("Warning: Failed to get latest block header: %v", err)
-		return 0, common.Hash{}, 0
-	}
-
-	blockNumber := header.Number.Uint64()
-	blockHash := header.Hash()
-
-	// Get the full block to count transactions
-	block, err := ethClient.BlockByNumber(ctx, header.Number)
-	if err != nil {
-		t.Logf("Warning: Failed to get full block: %v", err)
-		t.Logf("Latest block: height=%d, hash=%s, txs=unknown", blockNumber, blockHash.Hex())
-		return blockNumber, blockHash, 0
-	}
-
-	txCount := len(block.Transactions())
-
-	//t.Logf("Latest block: height=%d, hash=%s, txs=%d", blockNumber, blockHash.Hex(), txCount)
-	return blockNumber, blockHash, txCount
-}
-
-func TestEngineExecution(t *testing.T) {
-	allPayloads := make([][]types.Tx, 0) // Slice to store payloads from build to sync phase
-
-	initialHeight := uint64(0)
-	genesisHash := common.HexToHash(GENESIS_HASH)
-	genesisTime := time.Now().UTC().Truncate(time.Second)
-	genesisStateRoot := common.HexToHash(GENESIS_STATEROOT)
-	rollkitGenesisStateRoot := types.Hash(genesisStateRoot[:])
-
-	t.Run("Build chain", func(tt *testing.T) {
-		jwtSecret := setupTestRethEngine(tt)
-
-		executionClient, err := NewPureEngineExecutionClient(
-			TEST_ETH_URL,
-			TEST_ENGINE_URL,
-			jwtSecret,
-			genesisHash,
-			common.Address{},
-		)
-		require.NoError(t, err)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		stateRoot, gasLimit, err := executionClient.InitChain(ctx, genesisTime, initialHeight, CHAIN_ID)
-		require.NoError(t, err)
-		require.Equal(t, rollkitGenesisStateRoot, stateRoot)
-		require.NotZero(t, gasLimit)
-
-		prevStateRoot := rollkitGenesisStateRoot
-		lastHeight, lastHash, lastTxs := checkLatestBlock(tt, ctx)
-
-		for blockHeight := initialHeight + 1; blockHeight <= 10; blockHeight++ {
-			nTxs := int(blockHeight) + 10
-			// randomly use no transactions
-			if blockHeight == 4 {
-				nTxs = 0
-			}
-			txs := make([]*ethTypes.Transaction, nTxs)
-			for i := range txs {
-				txs[i] = getRandomTransaction(t, 22000)
-			}
-			for i := range txs {
-				submitTransaction(t, txs[i])
-			}
-			time.Sleep(1000 * time.Millisecond)
-
-			payload, err := executionClient.GetTxs(ctx)
-			require.NoError(tt, err)
-			require.Len(tt, payload, nTxs+1)
-
-			allPayloads = append(allPayloads, payload)
-
-			// Check latest block before execution
-			beforeHeight, beforeHash, beforeTxs := checkLatestBlock(tt, ctx)
-			require.Equal(tt, lastHeight, beforeHeight, "Latest block height should match")
-			require.Equal(tt, lastHash.Hex(), beforeHash.Hex(), "Latest block hash should match")
-			require.Equal(tt, lastTxs, beforeTxs, "Number of transactions should match")
-
-			newStateRoot, maxBytes, err := executionClient.ExecuteTxs(ctx, payload, blockHeight, genesisTime, prevStateRoot)
-			require.NoError(tt, err)
-			require.NotZero(tt, maxBytes)
-
-			err = executionClient.SetFinal(ctx, blockHeight)
-			require.NoError(tt, err)
-
-			// Check latest block after execution
-			lastHeight, lastHash, lastTxs = checkLatestBlock(tt, ctx)
-			require.Equal(tt, blockHeight, lastHeight, "Latest block height should match")
-			require.NotEmpty(tt, lastHash.Hex(), "Latest block hash should not be empty")
-			require.GreaterOrEqual(tt, lastTxs, 0, "Number of transactions should be non-negative")
-
-			if nTxs == 0 {
-				require.Equal(tt, prevStateRoot, newStateRoot)
-			} else {
-				require.NotEqual(tt, prevStateRoot, newStateRoot)
-			}
-			prevStateRoot = newStateRoot
-		}
-	})
-
-	// start new container and try to sync
-	t.Run("Sync chain", func(tt *testing.T) {
-		jwtSecret := setupTestRethEngine(t)
-
-		executionClient, err := NewPureEngineExecutionClient(
-			TEST_ETH_URL,
-			TEST_ENGINE_URL,
-			jwtSecret,
-			genesisHash,
-			common.Address{},
-		)
-		require.NoError(t, err)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		stateRoot, gasLimit, err := executionClient.InitChain(ctx, genesisTime, initialHeight, CHAIN_ID)
-		require.NoError(t, err)
-		require.Equal(t, rollkitGenesisStateRoot, stateRoot)
-		require.NotZero(t, gasLimit)
-
-		prevStateRoot := rollkitGenesisStateRoot
-		lastHeight, lastHash, lastTxs := checkLatestBlock(tt, ctx)
-
-		for blockHeight := initialHeight + 1; blockHeight-initialHeight <= 10; blockHeight++ {
-			payload := allPayloads[blockHeight-initialHeight-1]
-
-			// Check latest block before execution
-			beforeHeight, beforeHash, beforeTxs := checkLatestBlock(tt, ctx)
-			require.Equal(tt, lastHeight, beforeHeight, "Latest block height should match")
-			require.Equal(tt, lastHash.Hex(), beforeHash.Hex(), "Latest block hash should match")
-			require.Equal(tt, lastTxs, beforeTxs, "Number of transactions should match")
-
-			newStateRoot, maxBytes, err := executionClient.ExecuteTxs(ctx, payload, blockHeight, genesisTime, prevStateRoot)
-			require.NoErrorf(tt, err, "blockHeight: %d, nTxs: %d", blockHeight, len(payload)-1)
-			require.NotZero(tt, maxBytes)
-
-			err = executionClient.SetFinal(ctx, blockHeight)
-			require.NoError(tt, err)
-
-			// Check latest block after execution
-			lastHeight, lastHash, lastTxs = checkLatestBlock(tt, ctx)
-			require.Equal(tt, blockHeight, lastHeight, "Latest block height should match")
-			require.NotEmpty(tt, lastHash.Hex(), "Latest block hash should not be empty")
-			require.GreaterOrEqual(tt, lastTxs, 0, "Number of transactions should be non-negative")
-
-			if len(payload)-1 == 0 {
-				require.Equal(tt, prevStateRoot, newStateRoot)
-			} else {
-				require.NotEqual(tt, prevStateRoot, newStateRoot)
-			}
-			prevStateRoot = newStateRoot
-		}
-	})
-}
